Criado repository do PostgreSQL e em memória, criado Container para a API, simplificado os endpoints Batch e centralizadas variáveis de ambiente
