+++ Dockerfile-jupter
@@ -1,0 +1,2 @@
+# Arquivo utilizado pelo MYBinder, dessa forma o tempo de build é menor
+FROM julio471/jupter-hadoop-pets:19.0
