+++ api-resources/run_vaccine_pipeline.sh
