+++ api/.env.example
@@ -1,4 +1,5 @@
-# Server Configuration
+# General Configuration
+RESOURCES_PATH='/api-resources'
 API_PORT=3000
 
 # Database Configuration
