+++ api/src/config/cloudinary.js
@@ -1,0 +1,8 @@
+module.exports = {
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+    folder: process.env.CLOUDINARY_UPLOAD_FOLDER,
+    format: process.env.CLOUDINARY_UPLOAD_FORMAT
+};
+
