+++ api/src/config/general.js
@@ -1,0 +1,5 @@
+module.exports = {
+    resourcesPath: process.env.RESOURCES_PATH || '/api-resources',
+    apiPort: process.env.API_PORT || 3000
+};
+
