+++ api/src/controllers/batch.controller.js
@@ -1,112 +1,43 @@
 const { dbConfig } = require('../config/database');
+const { resourcesPath } = require('../config/general');
 const batchService = require('../services/batch.service');
 
-exports.startVaccineRecommendationJob = async (req, res) => {
+const entityScriptMapping = {
+    'vaccine-recommendation': 'run_vaccine_pipeline.sh',
+    'booking-reference': 'run_booking_reference_pipeline.sh',
+    'booking-recommendation': 'run_booking_recommendation_pipeline.sh',
+    'ltv-by-pet-profile': 'run_ltv_by_pet_profile_pipeline.sh',
+};
+
+exports.startJob = async (req, res) => {
     try {
-        const command = `bash resources/scripts/run_vaccine_pipeline.sh ${dbConfig.user} ${dbConfig.host} ${dbConfig.database} ${dbConfig.password} ${dbConfig.port}`;
-        const result = await batchService.startJob('vaccine_recommendation', command);
+        const { entity } = req.params;
+        const script = entityScriptMapping[entity];
+        if (!script) {
+            return res.status(400).send({ message: 'Invalid entity.' });
+        }
+        const command = `bash ${resourcesPath}/${script} ${dbConfig.user} ${dbConfig.host} ${dbConfig.database} ${dbConfig.password} ${dbConfig.port}`;
+        const result = await batchService.startJob(entity, command);
         res.send(result);
     } catch (error) {
         res.status(400).send({ message: error.message });
     }
 };
 
-exports.getVaccineRecommendationJobStatus = async (req, res) => {
+exports.getJobStatus = async (req, res) => {
     try {
-        const result = await batchService.getJobStatus('vaccine_recommendation');
+        const { id } = req.params;
+        const result = await batchService.getJobStatus(id);
         res.send(result);
     } catch (error) {
         res.status(400).send({ message: error.message });
     }
 };
 
-exports.getVaccineRecommendationJobResult = async (req, res) => {
+exports.getJobResult = async (req, res) => {
     try {
-        const result = await batchService.getJobResult('vaccine_recommendation');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.startBookingReferenceJob = async (req, res) => {
-    try {
-        const command = `bash resources/scripts/run_booking_reference_pipeline.sh ${dbConfig.user} ${dbConfig.host} ${dbConfig.database} ${dbConfig.password} ${dbConfig.port}`;
-        const result = await batchService.startJob('booking_reference', command);
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getBookingReferenceJobStatus = async (req, res) => {
-    try {
-        const result = await batchService.getJobStatus('booking_reference');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getBookingReferenceJobResult = async (req, res) => {
-    try {
-        const result = await batchService.getJobResult('booking_reference');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.startBookingRecommendationJob = async (req, res) => {
-    try {
-        const command = `bash resources/scripts/run_booking_recommendation_pipeline.sh ${dbConfig.user} ${dbConfig.host} ${dbConfig.database} ${dbConfig.password} ${dbConfig.port}`;
-        const result = await batchService.startJob('booking_recommendation', command);
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getBookingRecommendationJobStatus = async (req, res) => {
-    try {
-        const result = await batchService.getJobStatus('booking_recommendation');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getBookingRecommendationJobResult = async (req, res) => {
-    try {
-        const result = await batchService.getJobResult('booking_recommendation');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.startLtvByPetProfileJob = async (req, res) => {
-    try {
-        const command = `bash resources/scripts/run_ltv_by_pet_profile_pipeline.sh ${dbConfig.user} ${dbConfig.host} ${dbConfig.database} ${dbConfig.password} ${dbConfig.port}`;
-        const result = await batchService.startJob('ltv_by_pet_profile', command);
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getLtvByPetProfileJobStatus = async (req, res) => {
-    try {
-        const result = await batchService.getJobStatus('ltv_by_pet_profile');
-        res.send(result);
-    } catch (error) {
-        res.status(400).send({ message: error.message });
-    }
-};
-
-exports.getLtvByPetProfileJobResult = async (req, res) => {
-    try {
-        const result = await batchService.getJobResult('ltv_by_pet_profile');
+        const { id } = req.params;
+        const result = await batchService.getJobResult(id);
         res.send(result);
     } catch (error) {
         res.status(400).send({ message: error.message });
