+++ api/src/providers/email/nodemailer.provider.js
@@ -1,18 +1,19 @@
 const nodemailer = require('nodemailer');
+const { host, port, secure, user, pass, from } = require('../../config/email');
 
 const transporter = nodemailer.createTransport({
-    host: process.env.EMAIL_HOST,
-    port: process.env.EMAIL_PORT,
-    secure: process.env.EMAIL_SECURE === 'true',
+    host,
+    port,
+    secure,
     auth: {
-        user: process.env.EMAIL_USER,
-        pass: process.env.EMAIL_PASS
+        user,
+        pass
     }
 });
 
 const sendEmail = async ({ to, subject, html }) => {
     const mailOptions = {
-        from: process.env.EMAIL_FROM,
+        from,
         to,
         subject,
         html
