+++ api/src/providers/storage/cloudinary.provider.js
@@ -1,17 +1,18 @@
 const cloudinary = require('cloudinary').v2;
 const { CloudinaryStorage } = require('multer-storage-cloudinary');
+const { cloud_name, api_key, api_secret, folder, format } = require('../../config/cloudinary');
 
 cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET,
+  cloud_name,
+  api_key,
+  api_secret,
 });
 
 const storage = new CloudinaryStorage({
   cloudinary: cloudinary,
   params: {
-    folder: process.env.CLOUDINARY_UPLOAD_FOLDER,
-    format: async (req, file) => process.env.CLOUDINARY_UPLOAD_FORMAT,
+    folder,
+    format: async (req, file) => format,
     public_id: (req, file) => file.originalname,
   },
 });
