+++ labs/mapreduce-jobs/vaccine-recommendation/pom.xml
@@ -14,7 +14,7 @@
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <hadoop.version>2.9.2</hadoop.version> <!-- Corresponde ao Dockerfile -->
+        <hadoop.version>2.9.2</hadoop.version>
     </properties>
 
     <dependencies>
